modern-data-stack/requirements.txt
duckdb>=1.0.0
python-dotenv>=1.0.0
supabase~=2.10.0